final code

+++ src/main/java/spring/airplane/restapi/repository/BookingItemRepository.java
@@ -3,5 +3,5 @@ package spring.airplane.restapi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import spring.airplane.restapi.entity.BookingItem;
 
-public interface BookingItemRepository extends JpaRepository<BookingItem, Long> {
+public interface BookingItemRepository extends JpaRepository<BookingItem, Integer> {
 }
