+++ src/main/java/spring/airplane/restapi/repository/CreditHistoryRepository.java
@@ -3,5 +3,5 @@ package spring.airplane.restapi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import spring.airplane.restapi.entity.CreditHistory;
 
-public interface CreditHistoryRepository extends JpaRepository<CreditHistory, Long> {
+public interface CreditHistoryRepository extends JpaRepository<CreditHistory, Integer> {
 }
